# Improvements to map unification