--- a/unit-tests/Test/Booster/Fixture.hs
+++ b/unit-tests/Test/Booster/Fixture.hs
@@ -108,14 +108,47 @@ testKMapDefinition =
             , concatSymbolName = "Lbl'Unds'TestKMap'Unds'"
             }
 
-emptyKMap, concreteKMapWithOneItem, symbolicKMapWithOneItem :: Term
+kmapKeySort, kmapElementSort, kmapSort :: Sort
+kmapKeySort = SortApp testKMapDefinition.keySortName []
+kmapElementSort = SortApp testKMapDefinition.elementSortName []
+kmapSort = SortApp testKMapDefinition.mapSortName []
+
+emptyKMap
+    , concreteKMapWithOneItem
+    , concreteKMapWithTwoItems
+    , concreteKMapWithOneItemAndRest
+    , symbolicKMapWithOneItem
+    , concreteAndSymbolicKMapWithTwoItems ::
+        Term
 emptyKMap = KMap testKMapDefinition [] Nothing
 concreteKMapWithOneItem =
     KMap
         testKMapDefinition
         [
-            ( [trm| \dv{SomeSort{}}("key")|]
-            , [trm| \dv{SomeSort{}}("value")|]
+            ( [trm| \dv{SortTestKMapKey{}}("key") |]
+            , [trm| \dv{SortTestKMapItem{}}("value") |]
+            )
+        ]
+        Nothing
+concreteKMapWithOneItemAndRest =
+    KMap
+        testKMapDefinition
+        [
+            ( [trm| \dv{SortTestKMapKey{}}("key") |]
+            , [trm| \dv{SortTestKMapItem{}}("value") |]
+            )
+        ]
+        (Just [trm| REST:SortTestKMap{}|])
+concreteKMapWithTwoItems =
+    KMap
+        testKMapDefinition
+        [
+            ( [trm| \dv{SortTestKMapKey{}}("key") |]
+            , [trm| \dv{SortTestKMapItem{}}("value") |]
+            )
+        ,
+            ( [trm| \dv{SortTestKMapKey{}}("key2") |]
+            , [trm| \dv{SortTestKMapItem{}}("value2") |]
             )
         ]
         Nothing
@@ -123,8 +156,21 @@ symbolicKMapWithOneItem =
     KMap
         testKMapDefinition
         [
-            ( [trm| \dv{SomeSort{}}("key")|]
-            , [trm| A:SomeSort|]
+            ( [trm| \dv{SortTestKMapKey{}}("key") |]
+            , [trm| A:SortTestKMapItem{} |]
+            )
+        ]
+        Nothing
+concreteAndSymbolicKMapWithTwoItems =
+    KMap
+        testKMapDefinition
+        [
+            ( [trm| \dv{SortTestKMapKey{}}("key") |]
+            , [trm| \dv{SortTestKMapItem{}}("value") |]
+            )
+        ,
+            ( [trm| A:SortTestKMapKey{}|]
+            , [trm| \dv{SortTestKMapItem{}}("value2") |]
             )
         ]
         Nothing