This fixes #319 and hopefully also #298 by newly allowing unification of:
* Two maps with multiple fully concrete keys and no "rest" map variable (previously we could only match multiple keys if there was also a "rest of the map" variable)
* Map with only concrete keys + "rest" variable with a map of mixed concrete/symbolic keys without a "rest" variable