@goodlyrottenapple 2023-11-08
Kevm performance:
```
 Test                                                                                      | sam-map-unification-improvements time | master time        | (sam-map-unification-improvements/master) time
-------------------------------------------------------------------------------------------+---------------------------------------+--------------------+------------------------------------------------
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/hkg/approve-spec.k]             | 70.66                                 | 76.6               | 0.9224543080939948
 src/tests/integration/test_prove.py::test_pyk_prove[erc20/ds/transfer-failure-1-b-spec.k] | 169.29                                | 160.45             | 1.055095045185416
 TOTAL                                                                                     | 239.95                                | 237.04999999999998 | 1.0122337059692048
```